additional testing items. icon and picture.

+++ code_igniter/application/views/theme-tango/v_add_groups.php
@@ -2,7 +2,11 @@
 $sortcolumn = 3; 
 if (count($query) > 1)
 {
-?>		
+?>
+
+
+		<?php #echo "<pre>"; print_r($query); echo "</pre>\n"; ?>
+
 <table cellspacing="1" class="tablesorter">
 	<thead>
 		<tr>
@@ -22,8 +26,13 @@ if (count($query) > 1)
 				<td><?php echo preg_replace("/(\w+) attribute/", "<i>$1</i> attribute", $key['group_description']); ?></td>
 				
 				<?php if ($key['activated']) { ?>
-					<td align="center"><?php echo "<a href=\"action_delete_group/" . $key['activated'] . "\"><img src='" . $image_path . "16_delete.png' alt='' title='' width='16'/></a>";?></td>
-					<td align="center">yes</td>
+					<?php if ($key['activated'] != '1') { ?>
+						<td align="center"><?php echo "<a href=\"action_delete_group/" . $key['activated'] . "\"><img src='" . $image_path . "16_delete.png' alt='' title='' width='16'/></a>";?></td>
+						<td align="center">yes</td>
+					<?php } else { ?>
+						<td align="center"></td>
+						<td align="center">yes</td>
+					<?php } ?>
 				<?php } else { ?>
 					<td align="center"><?php echo "<a href=\"action_activate_group/" . $key['file'] . "\"><img src='" . $image_path . "16_true.png' alt='' title='' width='16'/></a>";?></td>
 					<td align="center"></td>
